refactor: :truck: renamed logo file name
