+++ new-website/src/components/logo/index.jsx
